requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.9
drf-spectacular==0.26.5
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
django-storages==1.14.2
google-cloud-storage==2.13.0
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
stripe==7.7.0
whitenoise==6.6.0
gunicorn==21.2.0
django-allauth==0.58.2
WeasyPrint==60.2
qrcode==7.4.2
django-debug-toolbar==4.2.0
paramiko>=2.12.0
requests>=2.28.0
openpyxl>=3.1.0
phonenumbers>=8.13.0